Time: 63 ms (37.67%), Space: 18.3 MB (62.91%) - LeetHub

+++ 115-distinct-subsequences/115-distinct-subsequences.cpp
@@ -22,7 +22,7 @@ public:
     // }
     int numDistinct(string s, string t) {
         int m = t.length(), n = s.length();
-        vector<vector<long long int>> dp(m + 1, vector<long long int> (n + 1, 0));
+        vector<vector<int>> dp(m + 1, vector<int> (n + 1, 0));
         for (int j = 0; j <= n; j++) dp[0][j] = 1;
         for (int j = 1; j <= n; j++)
             for (int i = 1; i <= m; i++)
